Add 'bacon-frying' badge to README.md
This adds a shield/badge to README.md with the text 'bacon-frying'.

The service provided by http://shields.io/ is used to generate the badge.

This parodies the badges used by many open-source projects on GitHub,
the image links to the Unfilter show page on jupiterbroadcasting.com
as a homage to the the origin of the phrase 'frying (conspiracy) bacon'.